02/01/2022

+++ src/Boletin_Tema7_Ejercicio7/Main_Fraccion.java
@@ -6,17 +6,17 @@ import java.util.Scanner;
 public class Main_Fraccion {
 
             static Scanner sc=new Scanner(System.in);
-            public static float pedirNumerador(){
+            public static int pedirNumerador(){
                 System.out.println("Introduzca el numerador");
-                return sc.nextFloat();
+                return sc.nextInt();
             }
 
-            public static float pedirDenominador() {
+            public static int pedirDenominador() {
 
-                float denominador=0;
+                int denominador=0;
                 do {
                     System.out.println("Introduzca el valor del denominador");
-                    denominador= sc.nextFloat();
+                    denominador= sc.nextInt();
                 } while (denominador <0);
                 return denominador;
             }
@@ -31,6 +31,7 @@ public class Main_Fraccion {
    // Fraccion fraccion2=new Fraccion(pedirNumerador(),pedirDenominador());
 
         System.out.println(  fraccion1.dividir(7));
+
